Module Single_value_abstraction.Log

Automatically log the calls to a basis in the trace.

module Sig : sig ... end

module Log_Numeric_Enum (Log : Tracelog.S) (Sub : Sig.NUMERIC_ENUM) : Sig.NUMERIC_ENUM

Logs all calls to Sub using Tracelog.

module No_Log_Numeric_Enum (Log : Tracelog.S) (Sub : Sig.NUMERIC_ENUM) : Sig.NUMERIC_ENUM

Does not logs all calls to Sub using Tracelog; directly use Sub.