Module WITH_MEMORY_BITVECTOR_BOOLEAN_FORWARD.Block_Forward

val sizeof : (block, bitvector) Operator.Forward_Arity.ar1

Size of a block in bytes

val concat : (block, block, block) Operator.Forward_Arity.ar2

Concatenates two blocks

val load : size:Units.In_bits.t -> (block, bitvector, bitvector) Operator.Forward_Arity.ar2

Loads (extracts) a value of a fixed size at a given index from a block

val store : size:Units.In_bits.t -> (block, bitvector, bitvector, block) Operator.Forward_Arity.ar3

Stores (writes) a fixed size value of a given index in a block

val binary_to_block : size:Units.In_bits.t -> (bitvector, block) Operator.Forward_Arity.ar1

Converts a fixed size value to a block